Modified progressBar.py so that the percentage it displays will have, at most, one decimal place.
The reason for this change is to insure the percentages look relatively uniform as the program is running.

Beforehand, when processing a batch of test cases, if a proportion of a
single test case to the total number of test cases resulted in a repeating
decimal, the percentage displayed on screen could be quite long. This ensures
a consistent look regardless of the number of test cases that are processed.